+++ source/packages/@aws-accelerator/accelerator/lib/stacks/network-stacks/network-vpc-endpoints-stack/network-vpc-endpoints-stack.ts
@@ -16,6 +16,7 @@ import { NagSuppressions } from 'cdk-nag';
 import { pascalCase } from 'change-case';
 import { Construct } from 'constructs';
 import * as path from 'path';
+import { getGlobalRegion } from '../../../../../../@aws-accelerator/utils/lib/common-functions';
 
 import {
   AseaResourceType,
@@ -728,10 +729,16 @@ export class NetworkVpcEndpointsStack extends NetworkStack {
     const privateDnsValue = !vpcItem.interfaceEndpoints?.central ?? true;
 
     for (const endpointItem of vpcItem.interfaceEndpoints?.endpoints ?? []) {
+      const globalRegion = getGlobalRegion(cdk.Stack.of(this).partition);
       if (this.isManagedByAsea(AseaResourceType.VPC_ENDPOINT, `${vpcItem.name}/${endpointItem.service}`)) {
         this.logger.info(
           `Interface Endpoint "${endpointItem.service}" for VPC "${vpcItem.name}" is managed externally`,
         );
+        continue;
+      }
+      // Skip creating hosted zone if service is 'iam'
+      if (endpointItem.service === 'iam' && cdk.Stack.of(this).region !== globalRegion) {
+        this.logger.info(`Skipping interface endpoint creation for IAM service in VPC: ${vpcItem.name}`);
         continue;
       }
       this.logger.info(`Adding Interface Endpoint for ${endpointItem.service}`);