+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix(asea): fix route53 query logging stack failures during upgrade
 - fix(security-audit): sns topic for controltower notifications
 - fix(macie): update macie custom resource policy statement to correct error preventing creation of service linked roler
+- fix(networking): allow the creation of iam vpc interface endpoints
 
 ### Changed